refactor(vote): Alteração de nomes de package
