--- api/src/main/java/com/dbserver/votingchallenge/domain/voting/VotingSessionRepository.java
@@ -1,6 +1,0 @@
-package com.dbserver.votingchallenge.domain.voting;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface VotingSessionRepository extends JpaRepository<VotingSession, Integer> {
-}
