+++ api/src/main/java/com/dbserver/votingchallenge/domain/votingSession/VotingSessionRepository.java
@@ -1,0 +1,10 @@
+package com.dbserver.votingchallenge.domain.votingSession;
+
+import com.dbserver.votingchallenge.domain.agenda.Agenda;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface VotingSessionRepository extends JpaRepository<VotingSession, Integer> {
+    Optional<VotingSession> findByAgenda(Agenda agenda);
+}
